backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
mangum==0.17.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
postgrest-py==0.13.0

# AI
google-generativeai==0.3.2

# Auth & Security
pyjwt==2.8.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Payments
stripe==7.8.0

# Email
resend==0.8.0

# Caching
upstash-redis==0.3.0

# Utils
python-dateutil==2.8.2
